+++ packages/compile/src/generate/gen-code-c.spec.ts
@@ -63,7 +63,87 @@ describe('generateC (Vensim -> C)', () => {
       y = :NOT: x ~~|
     `
     const code = readInlineModelAndGenerateC(mdl)
-    expect(code).toMatch('#include "sde.h"')
+    expect(code).toMatch(`\
+#include "sde.h"
+
+// Model variables
+double _x;
+double _y;`)
+  })
+
+  it('should work when valid input variable name without subscript is provided in spec file', () => {
+    const mdl = `
+      x = 10 ~~|
+      y = x + 1 ~~|
+    `
+    const code = readInlineModelAndGenerateC(mdl, {
+      inputVarNames: ['x'],
+      outputVarNames: ['y']
+    })
+    expect(code).toMatch(`\
+#include "sde.h"
+
+// Model variables
+double _x;
+double _y;`)
+  })
+
+  it('should work when valid input variable name with subscript (referenced by output variable) is provided in spec file', () => {
+    const mdl = `
+      DimA: A1, A2 ~~|
+      A[DimA] = 10, 20 ~~|
+      B[DimA] = A[DimA] + 1 ~~|
+    `
+    const code = readInlineModelAndGenerateC(mdl, {
+      inputVarNames: ['A[A1]'],
+      outputVarNames: ['B[A1]', 'B[A2]']
+    })
+    expect(code).toMatch(`\
+#include "sde.h"
+
+// Model variables
+double _a[2];
+double _b[2];`)
+  })
+
+  it('should work when valid input variable name with subscript (not referenced by output variable) is provided in spec file', () => {
+    // Note that `A` is specified as an input variable, but `A` is not referenced by output
+    // variable `B`, which is an unusual (but valid) usage scenario, so `A` should not be
+    // pruned by the `removeUnusedVariables` code (see #438)
+    const mdl = `
+      DimA: A1, A2 ~~|
+      A[DimA] = 10, 20 ~~|
+      B[DimA] = 30, 40 ~~|
+    `
+    const code = readInlineModelAndGenerateC(mdl, {
+      inputVarNames: ['A[A1]'],
+      outputVarNames: ['B[A1]', 'B[A2]']
+    })
+    expect(code).toMatch(`\
+#include "sde.h"
+
+// Model variables
+double _a[2];
+double _b[2];`)
+  })
+
+  // Note that this test takes a while (> 8 seconds) to run, so is skipped by default
+  // TODO: This test has little to do with code gen; we should move this and other similar
+  // tests to a separate file focused on the `removeUnusedVariables` step
+  it.skip('should work without exceeding stack limits when model has deep dependency tree', () => {
+    const n = 50000
+    let mdl = 'x0 = 1 ~~|\n'
+    for (let i = 1; i <= n; i++) {
+      mdl += `x${i} = x${i - 1} + 1 ~~|`
+    }
+
+    const code = readInlineModelAndGenerateC(mdl, {
+      inputVarNames: ['x0'],
+      outputVarNames: [`x${n}`]
+    })
+    for (let i = 0; i <= n; i++) {
+      expect(code).toContain(`double _x${i};`)
+    }
   })
 
   it('should throw error when unknown input variable name is provided in spec file', () => {