+++ packages/compile/src/parse-and-generate.js
@@ -147,7 +147,7 @@ export function printNames(namesPathname, operation) {
  * @return {*} A parsed tree representation of the model.
  */
 export function parseModel(input, modelDir, options) {
-  if (process.env.SDE_NONPUBLIC_USE_NEW_PARSE !== '1') {
+  if (process.env.SDE_NONPUBLIC_USE_NEW_PARSE === '0') {
     // Use the legacy parser
     return {
       kind: 'vensim-legacy',