+++ packages/create/CHANGELOG.md
@@ -1,5 +1,55 @@
 # Changelog
 
+## [0.2.14](https://github.com/climateinteractive/SDEverywhere/compare/create-v0.2.13...create-v0.2.14) (2024-04-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @sdeverywhere/compile bumped from ^0.7.16 to ^0.7.17
+
+## [0.2.13](https://github.com/climateinteractive/SDEverywhere/compare/create-v0.2.12...create-v0.2.13) (2024-04-19)
+
+
+### Bug Fixes
+
+* replace degit with giget so that `create` package doesn't depend on Git being installed locally ([#467](https://github.com/climateinteractive/SDEverywhere/issues/467)) ([0bf481b](https://github.com/climateinteractive/SDEverywhere/commit/0bf481b005ccb72075a186020098529be4004f9f)), closes [#464](https://github.com/climateinteractive/SDEverywhere/issues/464)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @sdeverywhere/compile bumped from ^0.7.15 to ^0.7.16
+
+## [0.2.12](https://github.com/climateinteractive/SDEverywhere/compare/create-v0.2.11...create-v0.2.12) (2024-04-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @sdeverywhere/compile bumped from ^0.7.14 to ^0.7.15
+
+## [0.2.11](https://github.com/climateinteractive/SDEverywhere/compare/create-v0.2.10...create-v0.2.11) (2024-03-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @sdeverywhere/compile bumped from ^0.7.13 to ^0.7.14
+
+## [0.2.10](https://github.com/climateinteractive/SDEverywhere/compare/create-v0.2.9...create-v0.2.10) (2024-03-11)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @sdeverywhere/compile bumped from ^0.7.12 to ^0.7.13
+
 ## [0.2.9](https://github.com/climateinteractive/SDEverywhere/compare/create-v0.2.8...create-v0.2.9) (2024-02-23)
 
 ### Dependencies